requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.18
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-celery==3.3.1
django-haystack==2.8.1
django-redis==4.12.1
django-redis-sessions==0.6.2
django-tinymce==2.8.0
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.8
importlib-metadata==3.4.0
itsdangerous==1.1.0
jieba==0.42.1
kombu==3.0.37
monotonic==1.5
mutagen==1.42.0
mysqlclient==2.0.3
Pillow==5.4.1
prompt-toolkit==3.0.11
py3Fdfs==2.2.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.24.2
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
zipp==3.4.0
